bsp: improved readability of functions, replaced floats by ints since this would cause some specific functions/registers of MIPS to be used; cpu: defining some specific instructions with macros and inline assembly; print_number.c: testing some specific instructions

+++ bsp.h
@@ -5,7 +5,7 @@ int read_w(int addr);
 void write_multiple(int src_addr,int dst_addr,int l);
 void write_vector(int src_addr,int dst_addr);
 
-float dot_product(int A_ptr,int B_ptr,uint l);
+int dot_product(int A_ptr,int B_ptr,int l);
 
 //constants
